+++ src/CategoriesCollection.php
@@ -9,14 +9,18 @@ use Nexendrie\Utils\Collection as BaseCollection;
  * CategoriesCollection
  *
  * @author Jakub Konečný
+ * @extends  BaseCollection<Category>
  */
 final class CategoriesCollection extends BaseCollection implements XmlConvertible
 {
-    protected string $class = Category::class;
+    public function __construct()
+    {
+        parent::__construct();
+        $this->class = Category::class;
+    }
 
     public function appendToXml(\SimpleXMLElement $parent): void
     {
-        // @phpstan-ignore argument.type
         array_walk($this->items, static function (Category $value) use ($parent): void {
             $value->appendToXml($parent);
         });